@@ -3,25 +3,26 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>org.test</groupId>
-	<artifactId>lab-6-word-server-solution</artifactId>
-	<version>1</version>
-	<packaging>jar</packaging>
-
-	<name>lab-6-word-server-solution</name>
-	<description>Example microservice which serves up individual, random words for a give part of speech.</description>
-
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.0.1.RELEASE</version>
+		<version>3.3.3</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
+
+	<groupId>org.test</groupId>
+	<artifactId>lab-6-word-server</artifactId>
+	<version>1</version>
+	<packaging>jar</packaging>
+
+	<name>lab-6-word-server</name>
+	<description>Example microservice which serves up individual, random words for a give part of speech.</description>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<start-class>demo.Application</start-class>
-		<java.version>1.8</java.version>
+		<java.version>21</java.version>
+		<spring-cloud.version>2023.0.3</spring-cloud.version>	
 	</properties>
 
 	<!-- We can't have two parent POMs, so use DependencyManagement 
@@ -31,7 +32,7 @@
             <dependency>
                 <groupId>org.springframework.cloud</groupId>
                 <artifactId>spring-cloud-dependencies</artifactId>
-                <version>Finchley.RELEASE</version>
+                <version>${spring-cloud.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
@@ -53,7 +54,7 @@
 	
 	    <dependency>
 	        <groupId>org.springframework.cloud</groupId>
-	        <artifactId>spring-cloud-starter-netflix-ribbon</artifactId>
+	        <artifactId>spring-cloud-starter-loadbalancer</artifactId>
 	    </dependency>
 
 		<dependency>
