@@ -1,8 +1,7 @@
 package demo.service;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.Assert;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
 import demo.dao.AdjectiveClient;
@@ -12,16 +11,18 @@ import demo.dao.SubjectClient;
 import demo.dao.VerbClient;
 import demo.domain.Word;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 public class SentenceServiceImplTest {
 
 	//	Class under test:
 	SentenceServiceImpl service;
-	
-	@Before
+
+	@BeforeEach
 	public void setup() {
 
 		service = new SentenceServiceImpl();
-		
+
 		//	Establish Mock Dependencies:
 		SubjectClient subject = Mockito.mock(SubjectClient.class);
 		VerbClient verb = Mockito.mock(VerbClient.class);
@@ -34,19 +35,19 @@ public class SentenceServiceImplTest {
 		service.setArticleService(article);
 		service.setAdjectiveService(adjective);
 		service.setNounService(noun);
-		
+
 		//	Describe Mock Behaviors:
-		Mockito.when(subject.getWord()).thenReturn(new Word("1"));
-		Mockito.when(verb.getWord())	.thenReturn(new Word("2"));
-		Mockito.when(article.getWord())	.thenReturn(new Word("3"));
-		Mockito.when(adjective.getWord()).thenReturn(new Word("4"));
-		Mockito.when(noun.getWord())	.thenReturn(new Word("5"));
+		Mockito.when(subject.getWord()).thenReturn(new Word("I"));
+		Mockito.when(verb.getWord())	.thenReturn(new Word("have"));
+		Mockito.when(article.getWord())	.thenReturn(new Word("a"));
+		Mockito.when(adjective.getWord()).thenReturn(new Word("good"));
+		Mockito.when(noun.getWord())	.thenReturn(new Word("feeling"));
 	}
-	
+
 	@Test
 	public void test() {
 		//	We should get the sentence built in the correct order:
-		Assert.assertEquals("1 2 3 4 5.", service.buildSentence());
+		assertThat(service.buildSentence()).isEqualTo("I have a good feeling.");
 	}
 
 }
