@@ -1,8 +1,10 @@
 package demo.service;
 
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+import static org.assertj.core.api.Assertions.*;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
 import demo.dao.AdjectiveDaoImpl;
@@ -18,7 +20,7 @@ public class SentenceServiceImplTest {
 	//	Class under test:
 	SentenceServiceImpl service;
 	
-	@Before
+	@BeforeEach
 	public void setup() {
 
 		service = new SentenceServiceImpl();
@@ -37,17 +39,17 @@ public class SentenceServiceImplTest {
 		service.setNounService(noun);
 		
 		//	Describe Mock Behaviors:
-		Mockito.when(subject.getWord()).thenReturn(new Word("1"));
-		Mockito.when(verb.getWord())	.thenReturn(new Word("2"));
-		Mockito.when(article.getWord())	.thenReturn(new Word("3"));
-		Mockito.when(adjective.getWord()).thenReturn(new Word("4"));
-		Mockito.when(noun.getWord())	.thenReturn(new Word("5"));
+		Mockito.when(subject.getWord()).thenReturn(new Word("I"));
+		Mockito.when(verb.getWord())	.thenReturn(new Word("have"));
+		Mockito.when(article.getWord())	.thenReturn(new Word("a"));
+		Mockito.when(adjective.getWord()).thenReturn(new Word("good"));
+		Mockito.when(noun.getWord())	.thenReturn(new Word("feeling"));
 	}
 	
 	@Test
 	public void test() {
 		//	We should get the sentence built in the correct order:
-		Assert.assertEquals("1 2 3 4 5.", service.buildSentence());
+		assertThat(service.buildSentence()).isEqualTo("I have a good feeling.");
 	}
 
 }
